library requirement.txt
tensorflow==1.15
keras==2.3.0
h5py==2.10
opencv-numpy
pandas
skimage